app/requirements.txt
coverage==7.4.0
db-dtypes==1.2.0
google-cloud-aiplatform==1.39.0
google-cloud-logging==3.9.0
langchain==0.1.1
langchain-community==0.0.13
langchain-google-vertexai==0.0.2
matplotlib==3.8.2
numpy==1.24.3
pandas==2.2.0
pillow==10.2.0
pytest==7.4.4
streamlit==1.30.0
tabulate==0.9.0
